requirements.txt
aiogram~=2.24
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.6.0
async-timeout==4.0.1
asyncpg==0.27.0
attrs==21.2.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==4.1.5
django-jazzmin==2.6.0
djangorestframework==3.14.0
environs==8.0.0
frozenlist==1.2.0
idna==3.3
marshmallow==3.14.1
multidict==5.2.0
packaging==23.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.28.2
sqlparse==0.4.3
typing_extensions==4.0.0
urllib3==1.26.14
yarl==1.7.2

pymongo~=4.3.3